Java web Application using servlet
it is a part of my study and I implement the concept of servlet where I create two HTML webpages Student registration and login. Firstly, when the user runs this application from the registration page, the user enters the data and clicks on submit to register. user is navigated to the servlet and data is stored in the database. later on the user can log in, if the user is already registered then after clicking on login the user is navigated to the servlet with the greeting message. otherwise, user is navigating to the servlet but with the message please register first.